+++ storybooks/main-storybook/src/2-foundation/2-2_typography.mdx
@@ -7,7 +7,7 @@ import {
   TypoTableSmall,
   TypoTableTitle,
   TypoTableXlarge,
-} from "../utils/typoTable.tsx";
+} from "../utils/typo-table.tsx";
 
 <Meta title="Foundation/Typography" />
