+++ storybooks/main-storybook/src/2-foundation/2-3_radius.mdx
@@ -1,7 +1,7 @@
 import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import { html } from "lit";
 
-import ChangeRoundedBase from "../changeRoundedBase";
+import ChangeRoundedBase from "../change-rounded-base";
 import Description from "../utils/description.tsx";
 import Logo from "../utils/logo.tsx";
 import Round from "../utils/round.tsx";
